refactor: common 컴포넌트 타입 정리
